Add esm data to package json to fix #908
