+++ eslint-rules/custom-eslint-rules.test.cjs
@@ -1,12 +1,12 @@
-const noDirectNullCheck = require('./no-direct-null-check')
-const noDirectUndefinedCheck = require('./no-direct-undefined-check')
-const noDirectArrayCheck = require('./no-direct-array-check')
-const noDirectIsFunctionCheck = require('./no-direct-function-check')
-const noDirectObjectCheck = require('./no-direct-object-check')
-const noDirectStringCheck = require('./no-direct-string-check')
-const noDirectDateCheck = require('./no-direct-date-check')
-const noDirectNumberCheck = require('./no-direct-number-check')
-const noDirectBooleanCheck = require('./no-direct-boolean-check')
+const noDirectNullCheck = require('./no-direct-null-check.cjs')
+const noDirectUndefinedCheck = require('./no-direct-undefined-check.cjs')
+const noDirectArrayCheck = require('./no-direct-array-check.cjs')
+const noDirectIsFunctionCheck = require('./no-direct-function-check.cjs')
+const noDirectObjectCheck = require('./no-direct-object-check.cjs')
+const noDirectStringCheck = require('./no-direct-string-check.cjs')
+const noDirectDateCheck = require('./no-direct-date-check.cjs')
+const noDirectNumberCheck = require('./no-direct-number-check.cjs')
+const noDirectBooleanCheck = require('./no-direct-boolean-check.cjs')
 
 const { RuleTester } = require('eslint')
