+++ eslint-rules/no-direct-array-check.cjs
