+++ eslint-rules/no-direct-boolean-check.cjs
