+++ eslint-rules/no-direct-document-check.cjs
