+++ eslint-rules/no-direct-file-check.cjs
