+++ eslint-rules/no-direct-function-check.cjs
