+++ eslint-rules/no-direct-number-check.cjs
