+++ eslint-rules/no-direct-object-check.cjs
