+++ eslint-rules/no-direct-string-check.cjs
