+++ eslint-rules/no-direct-undefined-check.cjs
