+++ react/package.json
@@ -22,6 +22,14 @@
     "main": "dist/umd/index.js",
     "module": "dist/esm/index.js",
     "types": "dist/types",
+    "type": "module",
+    "exports": {
+        ".": {
+            "types": "./dist/types/index.d.ts",
+            "import": "./dist/esm/index.js",
+            "require": "./dist/umd/index.js"
+        }
+    },
     "files": [
         "dist/*",
         "README.md"
@@ -68,5 +76,8 @@
     },
     "dependencies": {
         "@sentry/types": "7.37.2"
+    },
+    "engines": {
+        "node": ">=20.1.0"
     }
 }
